requirements.txt
anyio==4.3.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
cachecontrol==0.14.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.1.1
cryptography==42.0.5
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
#decouple==0.0.7
defusedxml==0.7.1
distlib==0.3.8
dj-database-url #==2.1.0
Django>=3.0,<4.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-heroku #==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.3
fonttools==4.50.0
fqdn==1.5.1
gunicorn #==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
importlib-metadata==7.1.0
importlib-resources==6.4.0
ipykernel==6.29.3
ipython==8.13.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.17
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-client==8.6.0
jupyter-console==6.6.3
jupyter-core==5.7.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==2.12.5
jupyter-server-terminals==0.5.3
jupyterlab==4.1.2
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.2
jupyterlab-widgets==3.0.10
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
mistune==3.0.2
msgpack==1.0.8
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.6.0
notebook==7.1.1
notebook-shim==0.2.4
numpy==1.24.4
oauthlib==3.2.2
overrides==7.7.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pkgutil-resolve-name==1.3.10
platformdirs==4.2.0
prometheus-client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
psycopg2==2.9.9
psycopg2-binary==2.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.8.2
python-decouple==3.1
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
requests==2.31.0
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.1
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
termcolor==2.4.0
terminado==0.18.0
threadpoolctl==3.4.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
traitlets==5.14.1
types-python-dateutil==2.9.0.20240316
typing-extensions==4.10.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
virtualenv==20.25.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
werkzeug==3.0.1
whitenoise==6.6.0
widgetsnbextension==4.0.10
wrapt==1.16.0
zipp==3.18.2
